Fix edit profile page validator
Group validation methods in a single class and prevent throwing exception from validator
